requirements.txt
Pillow
PyAudio==0.2.11
PyMySQL
PyQt5
cmake
google-auth-oauthlib
google-auth-httplib2
httplib2
google-api-python-client
pickles
mysql-connector-python
scikit-learn
argparse
pandas
opencv-python
numpy
dlib
face-recognition